Move template package out of internal
Also introduce a changable LookupFunc which makes it easier to use
templatectl in tests and fake the env without touching the env of the
running system.
